src/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile src/python/requirements.in
#
ansible==9.2.0
    # via -r src/python/requirements.in
ansible-core==2.16.3
    # via ansible
attrs==23.2.0
    # via cmd2
autopage==0.5.2
    # via cliff
cachetools==5.3.2
    # via google-auth
certifi==2024.2.2
    # via
    #   kubernetes
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cliff==4.5.0
    # via
    #   osc-lib
    #   python-openstackclient
cmd2==2.4.3
    # via cliff
cryptography==42.0.3
    # via
    #   ansible-core
    #   openstacksdk
    #   python-openstackclient
debtcollector==2.5.0
    # via
    #   oslo-config
    #   oslo-utils
    #   python-keystoneclient
decorator==5.1.1
    # via
    #   dogpile-cache
    #   openstacksdk
dogpile-cache==1.3.1
    # via openstacksdk
google-auth==2.28.0
    # via kubernetes
idna==3.6
    # via requests
iso8601==2.1.0
    # via
    #   keystoneauth1
    #   openstacksdk
    #   oslo-utils
    #   python-novaclient
    #   python-openstackclient
jinja2==3.1.3
    # via ansible-core
jmespath==1.0.1
    # via openstacksdk
jsonpatch==1.33
    # via openstacksdk
jsonpointer==2.4
    # via jsonpatch
keystoneauth1==5.5.0
    # via
    #   openstacksdk
    #   osc-lib
    #   python-cinderclient
    #   python-keystoneclient
    #   python-novaclient
kubernetes==29.0.0
    # via -r src/python/requirements.in
markupsafe==2.1.5
    # via jinja2
msgpack==1.0.7
    # via oslo-serialization
netaddr==1.2.1
    # via
    #   oslo-config
    #   oslo-utils
netifaces==0.11.0
    # via
    #   openstacksdk
    #   oslo-utils
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
openstacksdk==2.1.0
    # via
    #   -r src/python/requirements.in
    #   osc-lib
    #   python-openstackclient
os-service-types==1.7.0
    # via
    #   keystoneauth1
    #   openstacksdk
osc-lib==3.0.0
    # via python-openstackclient
oslo-config==9.3.0
    # via python-keystoneclient
oslo-i18n==6.2.0
    # via
    #   osc-lib
    #   oslo-config
    #   oslo-utils
    #   python-cinderclient
    #   python-keystoneclient
    #   python-novaclient
    #   python-openstackclient
oslo-serialization==5.3.0
    # via
    #   python-keystoneclient
    #   python-novaclient
oslo-utils==7.0.0
    # via
    #   osc-lib
    #   oslo-serialization
    #   python-cinderclient
    #   python-keystoneclient
    #   python-novaclient
packaging==23.2
    # via
    #   ansible-core
    #   oslo-utils
    #   python-keystoneclient
pbr==6.0.0
    # via
    #   keystoneauth1
    #   openstacksdk
    #   os-service-types
    #   osc-lib
    #   oslo-i18n
    #   oslo-serialization
    #   python-cinderclient
    #   python-keystoneclient
    #   python-novaclient
    #   python-openstackclient
    #   stevedore
platformdirs==4.2.0
    # via openstacksdk
prettytable==3.10.0
    # via
    #   cliff
    #   python-cinderclient
    #   python-novaclient
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pyparsing==3.1.1
    # via oslo-utils
pyperclip==1.8.2
    # via cmd2
python-cinderclient==9.4.0
    # via python-openstackclient
python-dateutil==2.8.2
    # via kubernetes
python-keystoneclient==5.3.0
    # via python-openstackclient
python-novaclient==18.4.0
    # via python-openstackclient
python-openstackclient==6.5.0
    # via -r src/python/requirements.in
pytz==2024.1
    # via
    #   oslo-serialization
    #   oslo-utils
pyyaml==6.0.1
    # via
    #   ansible-core
    #   cliff
    #   kubernetes
    #   openstacksdk
    #   oslo-config
    #   oslo-utils
requests==2.31.0
    # via
    #   keystoneauth1
    #   kubernetes
    #   oslo-config
    #   python-cinderclient
    #   python-keystoneclient
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
requestsexceptions==1.4.0
    # via openstacksdk
resolvelib==1.0.1
    # via ansible-core
rfc3986==2.0.0
    # via oslo-config
rsa==4.9
    # via google-auth
simplejson==3.19.2
    # via osc-lib
six==1.16.0
    # via
    #   kubernetes
    #   python-dateutil
    #   python-keystoneclient
stevedore==5.1.0
    # via
    #   cliff
    #   dogpile-cache
    #   keystoneauth1
    #   osc-lib
    #   oslo-config
    #   python-cinderclient
    #   python-keystoneclient
    #   python-novaclient
    #   python-openstackclient
tzdata==2024.1
    # via
    #   oslo-serialization
    #   oslo-utils
urllib3==2.2.1
    # via
    #   kubernetes
    #   requests
wcwidth==0.2.13
    # via
    #   cmd2
    #   prettytable
websocket-client==1.7.0
    # via kubernetes
wrapt==1.16.0
    # via debtcollector
